// app/src/engine/types/Pesukim.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Pasuk } from "./Pasuk";
import type { PesukimLook } from "./PesukimLook";
import type { PesukimPlace } from "./PesukimPlace";

/**
 * Everything a Pesukim input holds.
 */
export type Pesukim = { 
/**
 * Always [`PESUKIM`] entries.
 */
pesukim: Array<Pasuk>, look: PesukimLook, place: PesukimPlace, };